+++ apps/ui-dev/layouts/BlankLayout.jsx
@@ -1,9 +1,12 @@
 import { View } from "@elf-framework/ui";
 
+import { useTheme } from "../hooks/useTheme";
 import "./BlankLayout.scss";
 
 export function BlankLayout(props) {
   const { content, maxWidth = 1200 } = props;
+
+  useTheme();
 
   return (
     <div class="blank-layout">